requirements-dev.txt
factory-boy==3.2.0
mkdocs==1.1.2
pytest==6.2.4
pytest-django==4.3.0
